+++ CHANGES.md
@@ -5,7 +5,8 @@ This files lists the changes during the lifetime of this project.
 - (backwards incompatible) clean slate - start over with only essential 4W functionality
 - (backwards incompatible) merge User and Colleague model
 - Change login to only pass when user is in database
-- Add first tests
+- Add authentication tests
+- Remove login requirement from logout endpoint
 
 ## 2025-10-09
 - add period filter to placement page